+++ include/Warcaby/Game/game.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "../Board/board.h"
-#include "../Player/player.h"
+#include "Warcaby/Board/board.h"
+#include "Warcaby/Player/player.h"
 
 /*
 Tylko funkcja Game ma mieć bezpośredni dostęp do klasy Board, inne funkcje
