+++ include/Warcaby/Player/player.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../types.h"
+#include "Warcaby/types.h"
 
 class Player{
     public:
